del_transfer/requirements.txt
pandas
mysql-connector-python
